server/requirements.txt
Django==6.0.3
djangorestframework==3.14.0
django-cors-headers==4.0.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
python-dotenv==1.0.0
Pillow==10.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
whitenoise==6.5.0
requests>=2.31.0
PyJWT[crypto]>=2.9.0
